Prediccion sin DH

+++ src/main/java/com/yarmovezzoli/gestioninv/DTOs/PrediccionDTO.java
@@ -13,7 +13,7 @@ import java.time.LocalDate;
 @NoArgsConstructor
 public class PrediccionDTO {
 
-    Double cantidadPredecida;
+    int cantidadPredecida;
     LocalDate fechaDesdePrediccion;
     LocalDate fechaHastaPrediccion;
     LocalDate fechaPrediccionRealizada;
